+++ batch-console-api/src/main/java/io/github/pinpols/batch/console/domain/governance/web/ConsoleGovernanceController.java
@@ -1,7 +1,7 @@
 package io.github.pinpols.batch.console.domain.governance.web;
 
 import io.github.pinpols.batch.common.dto.CommonResponse;
-import io.github.pinpols.batch.console.domain.observability.service.ConsoleSystemParameterService;
+import io.github.pinpols.batch.console.application.observability.ConsoleSystemParameterService;
 import io.github.pinpols.batch.console.service.ConsoleResponseFactory;
 import io.github.pinpols.batch.console.shared.query.TenantIdResolver;
 import io.github.pinpols.batch.console.support.web.ConsoleRequestMetadataResolver;