+++ batch-console-api/src/main/java/io/github/pinpols/batch/console/domain/workflow/web/ConsoleWorkflowDefinitionController.java
@@ -3,7 +3,6 @@ package io.github.pinpols.batch.console.domain.workflow.web;
 import io.github.pinpols.batch.common.dto.CommonResponse;
 import io.github.pinpols.batch.common.enums.ResultCode;
 import io.github.pinpols.batch.common.exception.BizException;
-import io.github.pinpols.batch.console.domain.job.web.request.EnabledPatchRequest;
 import io.github.pinpols.batch.console.domain.workflow.application.ConsoleWorkflowDefinitionApplicationService;
 import io.github.pinpols.batch.console.domain.workflow.application.ConsoleWorkflowDefinitionApplicationService.DagValidationResult;
 import io.github.pinpols.batch.console.domain.workflow.application.WorkflowDesignLockService;
@@ -16,6 +15,7 @@ import io.github.pinpols.batch.console.domain.workflow.web.response.WorkflowDefi
 import io.github.pinpols.batch.console.domain.workflow.web.response.WorkflowDesignLockResponse;
 import io.github.pinpols.batch.console.domain.workflow.web.response.WorkflowMermaidResponse;
 import io.github.pinpols.batch.console.service.ConsoleResponseFactory;
+import io.github.pinpols.batch.console.shared.command.EnabledPatchRequest;
 import io.github.pinpols.batch.console.shared.security.ConsolePrincipal;
 import io.github.pinpols.batch.console.support.web.Idempotent;
 import jakarta.validation.Valid;